+++ src/game/camera/components.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::Component;
+
+#[derive(Component)]
+pub struct GameCamera;
